refactor: パッケージの凝集度を改善
